requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
streamlit_lottie
tabulate
plotly.express
geopandas
joblib
scikit-learn
folium
streamlit_folium
pydeck